requirements_new.txt
absl-py==0.7.1
argon2-cffi==20.1.0
-e git+https://github.com/argoai/argoverse-api.git@9c9fb20e2651e7164ded706a27a70df04708c885#egg=argoverse
asciitree==0.3.3
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bleach==3.2.1
Bottleneck==1.2.1
cached-property==1.5.2
cachetools==4.1.1
cairocffi==1.2.0
CairoSVG==2.5.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colour==0.1.5
cssselect2==0.4.1
cycler==0.10.0
Cython==0.29.16
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
descartes==1.1.0
docutils==0.16
entrypoints==0.3
fasteners==0.15
filelock==3.0.12
fsspec==0.8.4
future==0.18.2
gast==0.3.3
gdown==3.12.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.1.8
grpcio==1.27.2
h5py==3.0.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib-metadata==2.0.0
ipdb==0.12.2
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.13.2
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Kivy==1.11.1
Kivy-Garden==0.1.4
kiwisolver==1.1.0
l5kit==1.1.0
lap==0.4.0
llvmlite==0.34.0
logger==1.4
lxml==4.6.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
monotonic==1.5
motmetrics==1.1.3
moviepy==1.0.3
mpmath==1.1.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
networkx==2.5
notebook==6.1.4
numba==0.51.2
numcodecs==0.7.2
numexpr==2.7.0
numpy==1.17.1
numpy-quaternion==2020.11.2.17.0.49
oauthlib==3.1.0
opencv-contrib-python-headless==4.4.0.44
opencv-python==4.4.0.44
opt-einsum==3.1.0
packaging==20.4
pandas==0.25.1
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
ply==3.11
proglog==0.1.9
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.13.0
PTable==0.9.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.2
pymap3d==2.4.3
pyntcloud @ git+https://github.com/daavoo/pyntcloud@60eb7d06bf46c040a7845a59702de4c86f012a50
pyparsing==2.4.2
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.0
pytorch-ignite==0.4.2
pytorch-lightning==1.0.4
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
Quaternion==3.5.2.post4
reportlab==3.5.55
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.21.3
scipy==1.5.3
seaborn==0.11.0
Send2Trash==1.5.0
Shapely==1.6.4.post2
six==1.14.0
sklearn==0.0
svglib==1.0.1
svgwrite==1.4
sympy==1.4
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.10.1
tinycss2==1.1.0
torch==1.7.0
torchvision==0.8.1
tornado==6.1
tqdm==4.51.0
traitlets==5.0.5
transforms3d==0.3.1
typing==3.7.4.1
typing-extensions==3.7.4.3
umap-learn==0.4.6
urllib3==1.25.11
virtualenv==16.7.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
xarray==0.14.0
zarr==2.5.0
zipp==3.4.0
